Refactoring: Updated 24 files, deleted 9 files
