+++ .env
@@ -1,8 +1,6 @@
 NEXT_PUBLIC_BUBE_API_ENDPOINT=https://1hmfpsvto6.execute-api.ap-northeast-1.amazonaws.com/dev
 NEXT_PUBLIC_BLOG_API_ENDPOINT=https://nsb2024.microcms.io/api/v1
 
-
-
 # This was inserted by `prisma init`:
 # Environment variables declared in this file are automatically made available to Prisma.
 # See the documentation for more detail: https://pris.ly/d/prisma-schema#accessing-environment-variables-from-the-schema
