+++ src/app/_components/elements/PageTitleHeader.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import React from "react";
 
 type Props = {
