+++ src/app/api/[invalid-path]/route.ts
@@ -5,6 +5,8 @@ import { StatusCodes } from "@/app/_utils/extendedStatusCodes";
 import { Origin } from "@/app/_types/ApiResponse";
 import AppErrorCode from "@/app/_types/AppErrorCode";
 
+// 存在しないエンドポイントへのリクエストを処理する
+// FIXME: /api/admin/hogehoge のような不正なパスは補足できない。
 // 共通処理
 const handleRequest = (req: NextRequest) => {
   const originalPath = req.nextUrl.pathname;
