+++ src/app/api/admin/posts/route.ts
@@ -8,43 +8,44 @@ import {
 import { StatusCodes } from "@/app/_utils/extendedStatusCodes";
 import SuccessResponseBuilder from "@/app/api/_helpers/successResponseBuilder";
 import ErrorResponseBuilder from "@/app/api/_helpers/errorResponseBuilder";
-import { GET as getPosts } from "@/app/api/posts/route";
 import { z } from "zod";
 import PostRequest from "@/app/_types/PostRequest";
 import PostService from "@/app/_services/postService";
 import AppErrorCode from "@/app/_types/AppErrorCode";
 
 // [GET] /api/admin/posts
-// NOTE:現状で /apy/posts と同じ処理
-export const GET = getPosts;
+export const GET = async () => {
+  try {
+    const posts = await PostService.fetchAllPosts();
+    return NextResponse.json(new SuccessResponseBuilder(posts).build());
+  } catch (error) {
+    const payload = new ErrorResponseBuilder().setUnknownError(error).build();
+    return NextResponse.json(payload, { status: payload.httpStatus });
+  }
+};
 
 // [POST] /api/admin/posts
 export const POST = async (req: NextRequest) => {
   try {
     const body: PostRequest.Payload = await req.json();
-
-    // バリデーションに問題があれば z.ZodError が Throw
     const validatedBody = PostRequest.serverValidationSchema.parse(body);
-
-    // DB関連で問題があればエラーが Throw
     const insertedPost = await PostService.insertPostWithCategories(
       validatedBody
     );
-
-    const payload = createSuccessResponse(insertedPost);
+    const payload = createPostSuccessResponse(insertedPost);
     return NextResponse.json(payload, { status: payload.httpStatus });
   } catch (error) {
-    const payload = createErrorResponse(error);
+    const payload = createPostErrorResponse(error);
     return NextResponse.json(payload, { status: payload.httpStatus });
   }
 };
 
-// 成功時のレスポンスを生成
-const createSuccessResponse = (post: Post): ApiSuccessResponse<Post> =>
+// [POST]成功時のレスポンス(201)を生成
+const createPostSuccessResponse = (post: Post): ApiSuccessResponse<Post> =>
   new SuccessResponseBuilder(post).setHttpStatus(StatusCodes.CREATED).build();
 
-// 失敗時のレスポンスを生成
-const createErrorResponse = (error: unknown): ApiErrorResponse => {
+// [POST]失敗時のレスポンスを生成
+const createPostErrorResponse = (error: unknown): ApiErrorResponse => {
   const errorResponseBuilder = new ErrorResponseBuilder();
   if (error instanceof z.ZodError) {
     const msg =
