--- src/app/posts/[id]/_types/GetPostResponse.ts
@@ -1,6 +1,0 @@
-import MicroCmsPost from "@/app/_types/MicroCmsPost";
-
-// ブログ記事（単体）の取得APIのレスポンス
-type GetPostResponse = MicroCmsPost;
-
-export default GetPostResponse;
