+++ src/app/posts/_components/PostSummary.tsx
@@ -2,21 +2,18 @@
 
 import Link from "next/link";
 import React from "react";
-import { formatIso8601ToJpDateTime } from "@/app/_utils/dateTimeUtils";
 import { format } from "date-fns";
 import DOMPurify from "isomorphic-dompurify";
-import Tag from "@/app/_components/elements/Tag";
-import MicroCmsPost from "@/app/_types/MicroCmsPost";
+import CategoryTag from "@/app/posts/_components/CategoryTag";
 import PostWithCategory from "@/app/admin/posts/_types/PostWithCategory";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faCalendarDays } from "@fortawesome/free-solid-svg-icons";
 
 type Props = {
-  // post: MicroCmsPost;
   post: PostWithCategory;
 };
 
-const ArticleSummary: React.FC<Props> = (props) => {
+const PostSummary: React.FC<Props> = (props) => {
   const { id, title, createdAt, categories, content } = props.post;
 
   const sanitizedContent: string = DOMPurify.sanitize(content, {
@@ -39,7 +36,7 @@ const ArticleSummary: React.FC<Props> = (props) => {
           {/* カテゴリ */}
           <div className="flex flex-wrap gap-1 items-center">
             {categories.map((c) => (
-              <Tag name={c.category.name} key={c.category.id} />
+              <CategoryTag name={c.category.name} key={c.category.id} />
             ))}
           </div>
         </div>
@@ -55,4 +52,4 @@ const ArticleSummary: React.FC<Props> = (props) => {
   );
 };
 
-export default ArticleSummary;
+export default PostSummary;
