+++ src/app/posts/_components/Posts.tsx
@@ -1,25 +1,19 @@
 "use client";
 
 import React from "react";
+import useGetRequest from "@/app/_hooks/useGetRequest";
+import PostWithCategory from "@/app/admin/posts/_types/PostWithCategory";
+import composeApiErrorMessage from "@/app/_utils/composeApiErrorMsg";
 import PostSummary from "./PostSummary";
 import FetchLoading from "@/app/_components/elements/FetchLoading";
 import FetchError from "@/app/_components/elements/FetchError";
-// import usePosts from "../_hooks/usePosts";
-
-import useGetRequest from "@/app/_hooks/useGetRequest";
-import PostWithCategory from "@/app/admin/posts/_types/PostWithCategory";
-import { ApiResponse } from "@/app/_types/ApiResponse";
-import composeApiErrorMessage from "@/app/_utils/composeApiErrorMsg";
 
 const Posts = () => {
-  // カスタムフック usePosts で記事一覧を取得
-  // const { data, error, isLoading, endpoint } = usePosts();
   const url = "/api/admin/posts";
   const { data, error } = useGetRequest<PostWithCategory[]>(url);
 
   // Fetch failed
   if (error) {
-    // return <FetchError apiEndpoint={endpoint} message={error.message} />;
     <FetchError apiEndpoint={url} message={composeApiErrorMessage(error)} />;
   }
 
@@ -32,10 +26,12 @@ const Posts = () => {
     return <div>記事は0件です。</div>;
   }
 
+  const posts = data.data as NonNullable<PostWithCategory[]>;
+
   return (
     <div className="mt-5 flex flex-col justify-center ">
-      {data.data?.map((content) => (
-        <PostSummary key={content.id} post={content} />
+      {posts.map((post) => (
+        <PostSummary key={post.id} post={post} />
       ))}
     </div>
   );
